+++ CHANGELOG.md
@@ -5,12 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.5.2] - 2023-04-28
+- fix #35 **begin()** in FRAM9 and FRAM11
+
+
 ## [0.5.2] - 2023-04-22
 - add **FRAM_ML** class - see FRAM_MULTILANGUAGE.md
 - add examples.
 - update readme.md.
 - update keywords.txt
-
 
 ## [0.5.1] - 2023-04-20
 - add **readUntil(uint16_t memaddr, char \* buf, uint16_t buflen, char separator)**.